+++ .gitattributes
@@ -1,0 +1,4 @@
+/gradlew  text eol=lf
+
+*.bat  text eol=crlf
+
