--- MANIFEST.MF
@@ -1,1 +1,0 @@
-Main-Class: Main
