+++ build.gradle
@@ -1,0 +1,18 @@
+apply plugin :'java'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+jar {
+    archiveBaseName = 'crepel'
+    archiveVersion ='0.1.0'
+
+    manifest{
+        attributes 'Main-Class' : "Main"
+    }
+}
+
+apply plugin : 'application'
+
+mainClassName = 'Main'
+
