+++ src/main/java/token/Span.java
